refactor(core): separate role-user related APIs from roles routes
