+++ packages/core/src/routes/role.ts
@@ -1,13 +1,7 @@
 import type { RoleResponse } from '@logto/schemas';
-import {
-  userInfoSelectFields,
-  userProfileResponseGuard,
-  Roles,
-  RoleType,
-  Users,
-} from '@logto/schemas';
+import { Roles, Users } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
-import { pick, tryThat } from '@silverhand/essentials';
+import { tryThat } from '@silverhand/essentials';
 import { object, string, z, number } from 'zod';
 
 import RequestError from '#src/errors/RequestError/index.js';
@@ -18,17 +12,14 @@ import koaRoleRlsErrorHandler from '#src/middleware/koa-role-rls-error-handler.j
 import assertThat from '#src/utils/assert-that.js';
 import { parseSearchParamsForSearch } from '#src/utils/search.js';
 
+import roleUserRoutes from './role.user.js';
 import type { AuthedRouter, RouterInitArgs } from './types.js';
 
-export default function roleRoutes<T extends AuthedRouter>(
-  ...[
-    router,
-    {
-      queries,
-      libraries: { quota },
-    },
-  ]: RouterInitArgs<T>
-) {
+export default function roleRoutes<T extends AuthedRouter>(...[router, tenant]: RouterInitArgs<T>) {
+  const {
+    queries,
+    libraries: { quota },
+  } = tenant;
   const {
     rolesScopes: { insertRolesScopes },
     roles: {
@@ -41,14 +32,8 @@ export default function roleRoutes<T extends AuthedRouter>(
       updateRoleById,
     },
     scopes: { findScopeById },
-    users: { findUserById, findUsersByIds, countUsers, findUsers },
-    usersRoles: {
-      countUsersRolesByRoleId,
-      deleteUsersRolesByUserIdAndRoleId,
-      findUsersRolesByRoleId,
-      findUsersRolesByUserId,
-      insertUsersRoles,
-    },
+    users: { findUsersByIds },
+    usersRoles: { countUsersRolesByRoleId, findUsersRolesByRoleId, findUsersRolesByUserId },
   } = queries;
 
   router.use('/roles(/.*)?', koaRoleRlsErrorHandler());
@@ -228,101 +213,5 @@ export default function roleRoutes<T extends AuthedRouter>(
     }
   );
 
-  router.get(
-    '/roles/:id/users',
-    koaPagination(),
-    koaGuard({
-      params: object({ id: string().min(1) }),
-      response: userProfileResponseGuard.array(),
-      status: [200, 400, 404],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { id },
-      } = ctx.guard;
-      const { limit, offset } = ctx.pagination;
-      const { searchParams } = ctx.request.URL;
-
-      await findRoleById(id);
-
-      return tryThat(
-        async () => {
-          const search = parseSearchParamsForSearch(searchParams);
-          const usersRoles = await findUsersRolesByRoleId(id);
-          const userIds = usersRoles.map(({ userId }) => userId);
-
-          const [{ count }, users] = await Promise.all([
-            countUsers(search, undefined, userIds),
-            findUsers(limit, offset, search, undefined, userIds),
-          ]);
-
-          ctx.pagination.totalCount = count;
-          ctx.body = users.map((user) => pick(user, ...userInfoSelectFields));
-
-          return next();
-        },
-        (error) => {
-          if (error instanceof TypeError) {
-            throw new RequestError(
-              { code: 'request.invalid_input', details: error.message },
-              error
-            );
-          }
-          throw error;
-        }
-      );
-    }
-  );
-
-  router.post(
-    '/roles/:id/users',
-    koaGuard({
-      params: object({ id: string().min(1) }),
-      body: object({ userIds: string().min(1).array().nonempty() }),
-      status: [201, 404, 422],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { id },
-        body: { userIds },
-      } = ctx.guard;
-
-      const role = await findRoleById(id);
-      assertThat(
-        role.type === RoleType.User,
-        new RequestError({ code: 'user.invalid_role_type', status: 422, roleId: role.id })
-      );
-
-      const usersRoles = await findUsersRolesByRoleId(id);
-      const existingUserIds = new Set(usersRoles.map(({ userId }) => userId));
-      const userIdsToAdd = userIds.filter((id) => !existingUserIds.has(id)); // Skip existing user ids.
-
-      if (userIdsToAdd.length > 0) {
-        await Promise.all(userIdsToAdd.map(async (userId) => findUserById(userId)));
-        await insertUsersRoles(
-          userIdsToAdd.map((userId) => ({ id: generateStandardId(), roleId: id, userId }))
-        );
-      }
-      ctx.status = 201;
-
-      return next();
-    }
-  );
-
-  router.delete(
-    '/roles/:id/users/:userId',
-    koaGuard({
-      params: object({ id: string().min(1), userId: string().min(1) }),
-      status: [204, 404],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { id, userId },
-      } = ctx.guard;
-      await deleteUsersRolesByUserIdAndRoleId(userId, id);
-      ctx.status = 204;
-
-      return next();
-    }
-  );
+  roleUserRoutes(router, tenant);
 }
